+++ app/classes/shared/main_controller.py
@@ -566,7 +566,6 @@ class Controller:
             name=data["name"],
             server_uuid=server_fs_uuid,
             server_dir=new_server_path,
-            backup_path=backup_path,
             server_command=server_command,
             server_file=server_file,
             server_log_file=log_location,
@@ -576,7 +575,7 @@ class Controller:
             server_host=monitoring_host,
             server_type=monitoring_type,
         )
-        self.management.set_backup_config(
+        self.management.add_default_backup_config(
             new_server_id,
             backup_path,
         )
@@ -722,7 +721,6 @@ class Controller:
             server_name,
             server_id,
             new_server_dir,
-            backup_path,
             server_command,
             server_jar,
             server_log_file,
@@ -776,7 +774,6 @@ class Controller:
             server_name,
             server_id,
             new_server_dir,
-            backup_path,
             server_command,
             server_exe,
             server_log_file,
@@ -821,7 +818,6 @@ class Controller:
             server_name,
             server_id,
             new_server_dir,
-            backup_path,
             server_command,
             server_exe,
             server_log_file,
@@ -869,7 +865,6 @@ class Controller:
             server_name,
             server_id,
             new_server_dir,
-            backup_path,
             server_command,
             server_exe,
             server_log_file,
@@ -893,16 +888,13 @@ class Controller:
     # **********************************************************************************
 
     def rename_backup_dir(self, old_server_id, new_server_id, new_uuid):
-        server_data = self.servers.get_server_data_by_id(old_server_id)
         server_obj = self.servers.get_server_obj(new_server_id)
-        old_bu_path = server_data["backup_path"]
         ServerPermsController.backup_role_swap(old_server_id, new_server_id)
-        backup_path = old_bu_path
+        backup_path = os.path.join(self.helper.backup_path, old_server_id)
         backup_path = Path(backup_path)
         backup_path_components = list(backup_path.parts)
         backup_path_components[-1] = new_uuid
         new_bu_path = pathlib.PurePath(os.path.join(*backup_path_components))
-        server_obj.backup_path = new_bu_path
         default_backup_dir = os.path.join(self.helper.backup_path, new_uuid)
         try:
             os.rmdir(default_backup_dir)
@@ -916,7 +908,6 @@ class Controller:
         name: str,
         server_uuid: str,
         server_dir: str,
-        backup_path: str,
         server_command: str,
         server_file: str,
         server_log_file: str,
@@ -931,7 +922,6 @@ class Controller:
             name,
             server_uuid,
             server_dir,
-            backup_path,
             server_command,
             server_file,
             server_log_file,
@@ -996,16 +986,16 @@ class Controller:
                             f"Unable to delete server files for server with ID: "
                             f"{server_id} with error logged: {e}"
                         )
-                    if Helpers.check_path_exists(
-                        self.servers.get_server_data_by_id(server_id)["backup_path"]
-                    ):
-                        FileHelpers.del_dirs(
-                            Helpers.get_os_understandable_path(
-                                self.servers.get_server_data_by_id(server_id)[
-                                    "backup_path"
-                                ]
+                    backup_configs = HelpersManagement.get_backups_by_server(
+                        server_id, True
+                    )
+                    for config in backup_configs:
+                        if Helpers.check_path_exists(config.backup_location):
+                            FileHelpers.del_dirs(
+                                Helpers.get_os_understandable_path(
+                                    config.backup_location
+                                )
                             )
-                        )
 
                 # Cleanup scheduled tasks
                 try: