+++ app/classes/shared/tasks.py
@@ -140,7 +140,7 @@ class TasksManager:
                         )
 
                 elif command == "backup_server":
-                    svr.a_backup_server()
+                    svr.server_backup_threader(cmd["action_id"])
 
                 elif command == "update_executable":
                     svr.jar_update()
@@ -240,6 +240,7 @@ class TasksManager:
                                         "system"
                                     ),
                                     "command": schedule.command,
+                                    "action_id": schedule.action_id,
                                 }
                             ],
                         )
@@ -268,6 +269,7 @@ class TasksManager:
                                         "system"
                                     ),
                                     "command": schedule.command,
+                                    "action_id": schedule.action_id,
                                 }
                             ],
                         )
@@ -284,6 +286,7 @@ class TasksManager:
                                         "system"
                                     ),
                                     "command": schedule.command,
+                                    "action_id": schedule.action_id,
                                 }
                             ],
                         )
@@ -303,6 +306,7 @@ class TasksManager:
                                         "system"
                                     ),
                                     "command": schedule.command,
+                                    "action_id": schedule.action_id,
                                 }
                             ],
                         )
@@ -337,6 +341,7 @@ class TasksManager:
             job_data["cron_string"],
             job_data["parent"],
             job_data["delay"],
+            job_data["action_id"],
         )
 
         # Checks to make sure some doofus didn't actually make the newly
@@ -367,6 +372,7 @@ class TasksManager:
                                     "system"
                                 ),
                                 "command": job_data["command"],
+                                "action_id": job_data["action_id"],
                             }
                         ],
                     )
@@ -393,6 +399,7 @@ class TasksManager:
                                     "system"
                                 ),
                                 "command": job_data["command"],
+                                "action_id": job_data["action_id"],
                             }
                         ],
                     )
@@ -409,6 +416,7 @@ class TasksManager:
                                     "system"
                                 ),
                                 "command": job_data["command"],
+                                "action_id": job_data["action_id"],
                             }
                         ],
                     )
@@ -428,6 +436,7 @@ class TasksManager:
                                     "system"
                                 ),
                                 "command": job_data["command"],
+                                "action_id": job_data["action_id"],
                             }
                         ],
                     )
@@ -520,6 +529,7 @@ class TasksManager:
                                     "system"
                                 ),
                                 "command": job_data["command"],
+                                "action_id": job_data["action_id"],
                             }
                         ],
                     )
@@ -543,6 +553,7 @@ class TasksManager:
                                     "system"
                                 ),
                                 "command": job_data["command"],
+                                "action_id": job_data["action_id"],
                             }
                         ],
                     )
@@ -559,6 +570,7 @@ class TasksManager:
                                     "system"
                                 ),
                                 "command": job_data["command"],
+                                "action_id": job_data["action_id"],
                             }
                         ],
                     )
@@ -578,6 +590,7 @@ class TasksManager:
                                     "system"
                                 ),
                                 "command": job_data["command"],
+                                "action_id": job_data["action_id"],
                             }
                         ],
                     )
@@ -653,6 +666,7 @@ class TasksManager:
                                             "system"
                                         ),
                                         "command": schedule.command,
+                                        "action_id": schedule.action_id,
                                     }
                                 ],
                             )