+++ app/classes/web/routes/api/servers/server/tasks/task/index.py
@@ -22,6 +22,9 @@ task_patch_schema = {
         "action": {
             "type": "string",
         },
+        "action_id": {
+            "type": "string",
+        },
         "interval": {"type": "integer"},
         "interval_type": {
             "type": "string",